backend/requirements.txt
pathway
fastapi
uvicorn
requests
chromadb
openai
streamlit
python-dotenv
